Quiesce a number of warnings in test programs (signed vs unsigned comparison, unused variables, etc.)
[git-p4: depot-paths = "//depot/projects/EVALUATION/Main/common/": change = 3285]

+++ src/test/test_be_io_logcabinet.cpp
@@ -87,7 +87,7 @@ doLogCabinetTests()
 
 	tr1::shared_ptr<LogSheet> ls;
 	string lsname;
-	for (int i = 0; i < LOGSHEETCOUNT; i++) {
+	for (size_t i = 0; i < LOGSHEETCOUNT; i++) {
 		ostringstream sbuf;
 		sbuf << "logsheet" << i;
 		lsname = sbuf.str();
